refactor: move types to a common dir
