+++ src/@types/article.ts
